Move benchmarks to a separate file
