+++ godis_test.go
@@ -136,51 +136,3 @@ func TestDelete(t *testing.T) {
 	}
 
 }
-
-// Benchmark for Set method
-func BenchmarkSet(b *testing.B) {
-	db := setUp()
-	b.ResetTimer()
-
-	// run the Set method b.N times
-	for n := 0; n < b.N; n++ {
-		db.Set("key", "value")
-	}
-}
-
-// Benchmark for Get method
-func BenchmarkGet(b *testing.B) {
-	db := setUp()
-	db.Set("key", "value")
-	b.ResetTimer()
-
-	// run the Get method b.N times
-	for n := 0; n < b.N; n++ {
-		db.Get("key")
-	}
-}
-
-// Benchmark for Exists method
-func BenchmarkExists(b *testing.B) {
-	db := setUp()
-	db.Set("key", "value")
-	b.ResetTimer()
-
-	// run the Exists method b.N times
-	for n := 0; n < b.N; n++ {
-		db.Exists("key")
-	}
-}
-
-// Benchmark for Del method
-func BenchmarkDel(b *testing.B) {
-	db := setUp()
-	b.ResetTimer()
-
-	// Set a key and Del method b.N times
-	for n := 0; n < b.N; n++ {
-		// Not an accurate measure as we can't Del a key without setting it
-		db.Set("key", "value")
-		db.Del("key")
-	}
-}
